pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "SAEGenRec"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "YRW" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "ipython",
    "jupyterlab",
    "matplotlib",
    "mkdocs",
    "notebook",
    "numpy",
    "pandas",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn",
    "torch",
    "transformers",
    "trl",
    "deepspeed",
    "faiss-cpu",
    "fire",
    "loguru",
    "sentence-transformers",
    "tqdm",
    "sae-lens",
    "Pillow",
    "requests",
]
requires-python = "~=3.12.0"


[tool.ruff]
line-length = 99
src = ["SAEGenRec"]
include = ["pyproject.toml", "SAEGenRec/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["SAEGenRec"]
force-sort-within-sections = true

